test-requirements.txt
odoo14-addon-medical_administration_practitioner@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_administration_practitioner
odoo14-addon-medical-terminology-sct@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_terminology_sct
odoo14-addon-medical-certify@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_certify
odoo14-addon-medical-administration-location@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_administration_location
odoo14-addon-medical-administration-location-stock@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_administration_location_stock
odoo14-addon-medical-medication@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_medication
odoo14-addon-medical-terminology-atc@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_terminology_atc
odoo14-addon-account-journal-inter-company@git+https://github.com/tegin/cb-addons.git@14.0#subdirectory=setup/account_journal_inter_company
odoo14-addon-pos-close-approval@git+https://github.com/tegin/cb-addons.git@14.0#subdirectory=setup/pos_close_approval
odoo14-addon-pos-multiple-sessions@git+https://github.com/tegin/cb-addons.git@14.0#subdirectory=setup/pos_multiple_sessions
odoo14-addon-safe-box@git+https://github.com/tegin/cb-addons.git@14.0#subdirectory=setup/safe_box
odoo14-addon-cb-sale-report-invoice@git+https://github.com/tegin/cb-addons.git@14.0#subdirectory=setup/cb_sale_report_invoice
odoo14-addon-pos-inter-company@git+https://github.com/tegin/cb-addons.git@14.0#subdirectory=setup/pos_inter_company
odoo14-addon-pos-safe-box@git+https://github.com/tegin/cb-addons.git@14.0#subdirectory=setup/pos_safe_box
odoo14-addon-product-nomenclature@git+https://github.com/tegin/cb-addons.git@14.0#subdirectory=setup/product_nomenclature
odoo14-addon-remote-report-to-printer-label@git+https://github.com/tegin/cb-addons.git@14.0#subdirectory=setup/remote_report_to_printer_label
odoo14-addon-sale-commission-cancel@git+https://github.com/tegin/cb-addons.git@14.0#subdirectory=setup/sale_commission_cancel
odoo14-addon-sale-third-party@git+https://github.com/tegin/cb-addons.git@14.0#subdirectory=setup/sale_third_party
odoo14-addon-sequence-parser@git+https://github.com/tegin/cb-addons.git@14.0#subdirectory=setup/sequence_parser
odoo14-addon-web-flagbox@git+https://github.com/tegin/cb-addons.git@14.0#subdirectory=setup/web_flagbox
odoo14-addon-cb-mgmtsystem-issue@git+https://github.com/tegin/cb-addons.git@14.0#subdirectory=setup/cb_mgmtsystem_issue
odoo14-addon-web-history-back@git+https://github.com/tegin/cb-addons.git@14.0#subdirectory=setup/web_history_back
odoo14-addon-medical-administration-center@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_administration_center
odoo14-addon-medical-administration-encounter@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_administration_encounter
odoo14-addon-medical-administration-encounter-careplan@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_administration_encounter_careplan
odoo14-addon-medical-administration-practitioner@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_administration_practitioner
odoo14-addon-medical-administration-practitioner-specialty@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_administration_practitioner_specialty
odoo14-addon-medical-base@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_base
odoo14-addon-medical-clinical-careplan@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_clinical_careplan
odoo14-addon-medical-clinical-procedure@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_clinical_procedure
odoo14-addon-medical-clinical-request-group@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_clinical_request_group
odoo14-addon-medical-diagnostic-report@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_diagnostic_report
odoo14-addon-medical-financial-coverage@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_financial_coverage
odoo14-addon-medical-medication-request@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_medication_request
odoo14-addon-medical-terminology@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_terminology
odoo14-addon-medical-workflow@git+https://github.com/tegin/medical-fhir.git@14.0#subdirectory=setup/medical_workflow
odoo14-addon-queue-management@git+https://github.com/tegin/kiwi.git@14.0#subdirectory=setup/queue_management
